Ran ccpcheck which led to some refactoring. The Data class was replaced with a Data namespace. globalMainWindow was replaced by a singleton MAINWINDOW (). Some changes to the constructors for vmNumber led to the discovery of some old code in several parts of the code base that were using outdated constructs. All functors based on std::function are now passed by const reference. Style changes. Some fixes.

+++ VivenciaManager3/textdb.h
@@ -47,7 +47,7 @@ protected:
 
 	bool m_open;
 	bool m_needsaving;
-	int m_headerSize;
+	int64_t m_headerSize;
 	uint m_buffersize;
 	TF_TYPE m_type;
 	QString m_filename;
@@ -79,7 +79,7 @@ public:
 	bool setFieldValue ( const QString& field_name, const QString& value );
 	bool setFieldValue ( const uint field_index, const QString& value );
 
-	inline int sectionCount () const { return cfgData.count (); }
+	inline uint sectionCount () const { return cfgData.count (); }
 
 protected:
 	bool loadData();
